refactor: split node renderer scroll listener
